% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{estimate.mu}
\alias{estimate.mu}
\title{Calculate the intercept}
\usage{
estimate.mu(yt, Pi, dt = 1)
}
\arguments{
\item{yt}{N times p matrix of observations of the process}

\item{Pi}{matrix Pi}

\item{dt}{timestep of the process}
}
\value{
p-dimensional vector of intercepts
}
\description{
This function calculates the intercept for a given matrix Pi and data matrix.
}
