% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{generate_Pi}
\alias{generate_Pi}
\title{Generate matrix Pi}
\usage{
generate_Pi(p, r, sparse = TRUE)
}
\arguments{
\item{p}{dimension of the matrix}

\item{r}{cointegration rank}

\item{sparse}{logical value indicating if the matrix P should have a sparse form}
}
\value{
pxp matrix Pi
}
\description{
This function generates a random matrix Pi that defines a cointegrated VEC model
with a specified cointegration rank through Jordan normal form Pi = P^{-1}*J*P.
}
