% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_alpha.R
\name{pen.alpha}
\alias{pen.alpha}
\title{Fit VEC model with penalty on alpha}
\usage{
pen.alpha(
  X,
  r,
  dt = 1,
  equal.penalty = F,
  n.lambda = 100,
  n.cv = 10,
  alpha = 1
)
}
\arguments{
\item{X}{Data matrix}

\item{r}{cointegration rank}

\item{dt}{time step}

\item{equal.penalty}{logical value indicating, whether the same penalty value lambda should be used for all rows of alpha}

\item{n.lambda}{number of penalty values lambda to be used}

\item{n.cv}{number of iterations of the crossvalidation procedures for picking optimal penalty value lambda}
}
\value{
A list containing the following components:
}
\description{
This function loads data matrix X and fits a cointegrated VEC model
with the cointegration rank r, where elements of the loading matrix
alpha are penalized with elastic net penalty.
}
\section{Slots}{

\describe{
\item{\code{ALPHA}}{estimated matrix alpha, recalculated per time unit (alpha/\code{dt})}

\item{\code{BETA}}{estimated matrix beta}

\item{\code{PI}}{\code{ALPHA*BETA}}

\item{\code{OMEGA}}{the covariance matrix of the random errors, recalculated per time unit (Omega/\code{dt})}

\item{\code{MU}}{estimated intercepts, recalculated per time unit (mu/\code{dt})}

\item{\code{lambda.lasso}}{the selected lasso penalty}

\item{\code{lambda.ridge}}{the selected ridge penalty}

\item{\code{lambda.seq}}{the sequence of penalty values}

\item{\code{ALPHA.seq}}{the sequence of alpha estimates for each penalty value in lambda.seq (not divided by \code{dt})}

\item{\code{PI.seq}}{the sequence of Pi estimates for each penalty value in lambda.seq (not divided by \code{dt})}
}}

