% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_beta.R
\name{pen.beta}
\alias{pen.beta}
\title{Fit VEC model with penalty on beta}
\usage{
pen.beta(
  X,
  r,
  max.iter = 10,
  conv = 10^-2,
  nlambda = 100,
  n.cv = 20,
  glmnetthresh = 1e-04,
  dt = 1,
  equal.penalty = F
)
}
\arguments{
\item{X}{Data matrix}

\item{r}{cointegration rank}

\item{max.iter}{maximum number of iterations of the algorithm}

\item{conv}{convergence parameter}

\item{nlambda}{number of penalty values lambda to be used}

\item{n.cv}{number of iterations of the crossvalidation procedures for picking optimal penalty value lambda}

\item{glmnetthresh}{tolerance parameter for the interior call of glmnet function}

\item{dt}{timestep}

\item{equal.penalty}{logical value indicating, whether the same penalty value lambda should be used for all columns of beta}
}
\value{
A list containing the following components:
}
\description{
This function loads data matrix X and fits a cointegrated VEC model
with the cointegration rank r, where elements of the cointegration matrix
beta are penalized with lasso penalty.
}
\section{Slots}{

\describe{
\item{\code{BETA}}{estimated matrix beta}

\item{\code{ALPHA}}{estimated matrix alpha, recalculated per time unit (alpha/\code{dt})}

\item{\code{OMEGA}}{the covariance matrix of the random errors, recalculated per time unit (Omega/\code{dt})}

\item{\code{PI}}{\code{ALPHA*BETA}}

\item{\code{MU}}{estimated intercepts, recalculated per time unit (mu/\code{dt})}

\item{\code{it}}{number of iterations}

\item{\code{value.obj}}{values of the objective functions throughout iterations}

\item{\code{Pi.it}}{estimate of Pi throughout iterations (vectorized form)}
}}

