% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_PI_rank.R
\name{pen.PI.rank}
\alias{pen.PI.rank}
\title{Fit VEC model with rank penalty}
\usage{
pen.PI.rank(
  X,
  n.lambda,
  lambda.min = 1e-12,
  dt = 1,
  crit = "CV",
  n.cv = 100,
  alpha = 0.01
)
}
\arguments{
\item{X}{Data matrix}

\item{n.lambda}{number of penalty values lambda to be used}

\item{lambda.min}{the smallest allowed penalty value}

\item{dt}{timestep}

\item{crit}{Criterion for choosing optimal penalty value, the options are
\code{"CV"} (crossvalidated error), \code{"AIC"} (Akaike information criterion),
\code{"BIC"} (Bayes information criterion) and \code{"HQ"} (Hannan-Quinn criterion).}

\item{n.cv}{number of repetitions of the crossvalidation procedure}

\item{alpha}{proportion of the ridge penalty}
}
\value{
A list containing the following components:
}
\description{
This function loads data matrix X and fits a VEC model,
where elements of the matrix Pi are penalized with rank penalty.
}
\section{Slots}{

\describe{
\item{\code{PI}}{estimated matrix Pi, recalculated per time unit (Pi/\code{dt})}

\item{\code{MU}}{estimated intercepts, recalculated per time unit (mu/\code{dt})}

\item{\code{OMEGA}}{the covariance matrix of the random errors, recalculated per time unit (Omega/\code{dt})}

\item{\code{lambda}}{the selected penalty value}

\item{\code{lambda.seq}}{the sequence of penalty values}

\item{\code{crit}}{the type of criterion to select optimal lambda}

\item{\code{crit.seq}}{the sequence of values of the criterion throughout iterations}

\item{\code{Pi.seq}}{the sequence of Pi estimates for each penalty value in lambda.seq (not divided by \code{dt})}
}}

