% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_PI_restricted.R
\name{pen.PI.restricted}
\alias{pen.PI.restricted}
\title{Fit VEC model with lasso penalty and restricted rank}
\usage{
pen.PI.restricted(
  X,
  n.lambda,
  lambda.min = 1e-12,
  lambda.max = 0,
  r,
  maxiter = 100,
  dt = 1,
  crit = "CV",
  w.auto = TRUE,
  n.cv = 100,
  q = 2,
  weights = matrix(1, nrow = ncol(X), ncol = ncol(X))
)
}
\arguments{
\item{X}{Data matrix}

\item{n.lambda}{number of penalty values lambda to be used}

\item{lambda.min}{the smallest allowed penalty value}

\item{lambda.max}{the largest allowed penalty value (not used, calculated automatically)}

\item{r}{rank of Pi}

\item{maxiter}{maximum number of iterations}

\item{dt}{timestep of the process}

\item{crit}{Criterion for choosing optimal penalty value, the options are
\code{"CV"} (crossvalidated error), \code{"AIC"} (Akaike information criterion),
\code{"BIC"} (Bayes information criterion) and \code{"HQ"} (Hannan-Quinn criterion).}

\item{w.auto}{logical value, whether the sequence of timesteps alpha_k of the algorithm should be determined by an in-built procedure}

\item{n.cv}{number of repetitions of the crossvalidation procedure}

\item{q}{timestep parameter when timesteps of the algorithm are chosen automatically}

\item{weights}{pxp matrix of weights (default: identical weights)}
}
\value{
A list containing the following components:
}
\description{
This function loads data matrix X and fits a VEC model,
where elements of the matrix Pi are penalized with lasso penalty
while at the same time the rank of Pi is restricted to a given value.
}
\section{Slots}{

\describe{
\item{\code{PI}}{estimated matrix Pi, recalculated per time unit (Pi/\code{dt})}

\item{\code{MU}}{estimated intercepts, recalculated per time unit (mu/\code{dt})}

\item{\code{OMEGA}}{the covariance matrix of the random errors, recalculated per time unit (Omega/\code{dt})}

\item{\code{lambda}}{the selected penalty value}

\item{\code{PI.iter}}{matrix with estimates of Pi throughout iterations of the algorithm, each row contains vectorized Pi}

\item{\code{objective.iter}}{values of the objective function throughout iterations}

\item{\code{w}}{the sequence of timesteps of the algorithm}

\item{\code{lambda.seq}}{the sequence of penalty values}

\item{\code{crit}}{the type of criterion to select optimal lambda}

\item{\code{crit.seq}}{the sequence of values of the criterion throughout iterations}

\item{\code{Pi.seq}}{the sequence of Pi estimates for each penalty value in lambda.seq (not divided by \code{dt})}
}}

