% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_PI_unrestricted.R
\name{pen.PI.unrestricted}
\alias{pen.PI.unrestricted}
\title{Fit VEC model with lasso penalty and restricted rank}
\usage{
pen.PI.unrestricted(
  yt,
  crit = c("fixed", "CV", "AIC", "BIC", "HQ"),
  lambda,
  alpha = 1,
  n.lambda = 100,
  n.cv = 10,
  thresh = 1e-12,
  maxit = 1e+06,
  dt = 1
)
}
\arguments{
\item{yt}{Data matrix}

\item{crit}{Criterion for choosing optimal penalty value, the options are
\code{"fixed"} (given lambda), \code{"CV"} (crossvalidated error), \code{"AIC"} (Akaike information criterion),
\code{"BIC"} (Bayes information criterion) and \code{"HQ"} (Hannan-Quinn criterion).}

\item{lambda}{given lasso penalty when \code{crit="fixed"}}

\item{alpha}{vector of ridge penalties}

\item{n.lambda}{number of penalty values lambda to be used}

\item{n.cv}{number of repetitions of the crossvalidation procedure}

\item{thresh}{convergence parameter}

\item{maxit}{maximum number of iterations}

\item{dt}{timestep of the process}
}
\value{
A list containing the following components:
}
\description{
This function loads data matrix X and fits a VEC model,
where elements of the matrix Pi are penalized with lasso penalty
while at the same time the rank of Pi is restricted to a given value.
}
\section{Slots}{

\describe{
\item{\code{PI}}{estimated matrix Pi, recalculated per time unit (Pi/\code{dt})}

\item{\code{MU}}{estimated intercepts, recalculated per time unit (mu/\code{dt})}

\item{\code{OMEGA}}{the covariance matrix of the random errors, recalculated per time unit (Omega/\code{dt})}

\item{\code{lambda}}{the selected lasso penalty value}

\item{\code{alpha}}{the selected ridge penalty value}

\item{\code{lambda.seq}}{the sequence of penalty values}

\item{\code{PI.seq}}{the sequence of Pi estimates for each penalty value in lambda.seq (not divided by \code{dt})}
}}

