% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_qr.R
\name{pen.qr}
\alias{pen.qr}
\title{Fit VEC model with penalty imposed on QR decomposition of Pi}
\usage{
pen.qr(
  X,
  crit = c("fixed", "CV", "AIC", "BIC", "HQ"),
  lambda,
  n.lambda = 100,
  n.cv = 10,
  thresh = 1e-12,
  maxit = 1e+06,
  dt = 1,
  psi = 2
)
}
\arguments{
\item{X}{Data matrix}

\item{crit}{Criterion for choosing optimal penalty value, the options are
\code{"fixed"} (given lambda), \code{"CV"} (crossvalidated error), \code{"AIC"} (Akaike information criterion),
\code{"BIC"} (Bayes information criterion) and \code{"HQ"} (Hannan-Quinn criterion).}

\item{lambda}{given penalty value when \code{crit="fixed"}}

\item{n.lambda}{number of penalty values lambda to be used}

\item{n.cv}{number of repetitions of the crossvalidation procedure}

\item{thresh}{convergence parameter}

\item{maxit}{maximum number of iterations}

\item{dt}{timestep of the process}

\item{psi}{power of the weights in the objective function}
}
\value{
A list containing the following components:
}
\description{
This function loads data matrix X and fits a VEC model,
where matrix Pi is factorized by QR decomposition and a row-lasso penalty
is imposed on the matrix R
}
\section{Slots}{

\describe{
\item{\code{PI}}{estimated matrix Pi, recalculated per time unit (Pi/\code{dt})}

\item{\code{MU}}{estimated intercepts, recalculated per time unit (mu/\code{dt})}

\item{\code{OMEGA}}{the covariance matrix of the random errors, recalculated per time unit (Omega/\code{dt})}

\item{\code{lambda}}{the selected lasso penalty value}

\item{\code{S}}{the matrix Q}

\item{\code{R}}{the matrix R}

\item{\code{lambda.seq}}{the sequence of penalty values}

\item{\code{PI.seq}}{the sequence of Pi estimates for each penalty value in lambda.seq (not divided by \code{dt})}

\item{\code{R.seq}}{the sequence of R estimates for each penalty value in lambda.seq}
}}

