% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{perf.metrics}
\alias{perf.metrics}
\title{Performance metrics of the estimator}
\usage{
perf.metrics(Pi.est, Pi.true, y1, x1, y2, x2)
}
\arguments{
\item{Pi.est}{estimate of Pi}

\item{Pi.true}{true matrix Pi}

\item{y1}{differences of the process (training sample)}

\item{x1}{lagged values of the process (training sample)}

\item{y2}{differences of the process (testing sample)}

\item{x2}{lagged values of the process (testing sample)}
}
\value{
p-dimensional vector of intercepts
}
\description{
This function calculates performance metrics of an estimator
with respect to the true matrix Pi and forecasting of independent data.
}
\section{Slots}{

\describe{
\item{\code{MSPE}}{mean square error of one step prediction of the process}

\item{\code{MSE}}{mean square error of Pi}

\item{\code{r.norm}}{l2 norm of Pi.est divided by l2 norm of Pi.true}

\item{\code{angle}}{principal angle of the space spanned by Pi.est and Pi.true}

\item{\code{rank}}{rank of Pi.est}

\item{\code{zeros}}{number of zero elements in Pi.est}

\item{\code{correct.zeros}}{proportion of correctly identified zeros in Pi.est}

\item{\code{correct.nonzeros}}{proportion of correctly identified nonzeros in Pi.est}

\item{\code{SSE}}{sum of squared residuals}
}}

