% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{PI.OLS}
\alias{PI.OLS}
\title{Fit VEC model with OLS}
\usage{
PI.OLS(Y, Z, dt = 1, intercept = T)
}
\arguments{
\item{Y}{matrix of first differences of the process}

\item{Z}{matrix of lagged values of the process}

\item{dt}{timestep of the process}

\item{intercept}{logical value indicating whether the model includes the intercept or not}
}
\value{
A list containing the following components:
}
\description{
This function loads data matrix X and fits a VEC model
using the ordinary least squares method.
}
\section{Slots}{

\describe{
\item{\code{PI}}{estimated matrix Pi, recalculated per time unit (Pi/\code{dt})}

\item{\code{MU}}{estimated intercepts, recalculated per time unit (mu/\code{dt})}

\item{\code{OMEGA}}{the covariance matrix of the random errors, recalculated per time unit (Omega/\code{dt})}
}}

