% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{PI_johansen}
\alias{PI_johansen}
\title{Fit VEC model with Johansen procedure}
\usage{
PI_johansen(
  Y = NULL,
  Z = NULL,
  r = 1,
  A = NULL,
  H = NULL,
  dt = 0.1,
  intercept = TRUE,
  normalize = TRUE
)
}
\arguments{
\item{Y}{matrix of first differences of the process}

\item{Z}{matrix of lagged values of the process}

\item{r}{cointegration rank}

\item{A}{matrix of linear restrictions on the space spanned by the loading matrix}

\item{H}{matrix of linear restriction on the space spanned by the cointegration matrix}

\item{dt}{timestep of the process}

\item{intercept}{logical value indicating whether the model includes the intercept or not}

\item{normalize}{logical value indicating whether the cointegration matrix should be normalized to contain a unit matrix as the upper rxr block}
}
\value{
A list containing the following components:
}
\description{
This function loads data matrix X and fits a VEC model
with a restricted cointegration rank using the likelihood-based approach
of Søren Johansen.
}
\section{Slots}{

\describe{
\item{\code{N}}{number of observations of the process}

\item{\code{p}}{dimension of the process}

\item{\code{r}}{cointegration rank}

\item{\code{alpha}}{estimate of the loading matrix, recalculated per time unit (alpha/\code{dt})}

\item{\code{beta}}{estimate of the cointegration matrix}

\item{\code{Psi}}{estimated intercepts, recalculated per time unit (mu/\code{dt})}

\item{\code{Omega}}{the covariance matrix of the random errors, recalculated per time unit (Omega/\code{dt})}

\item{\code{test}}{test statistics of the trace test}

\item{\code{lambda}}{eigenvalues resulting from calculating the estimate of the cointegration matrix beta}

\item{\code{A}}{linear restrictions imposed on the space spanned by the loading matrix alpha}

\item{\code{H}}{linear restrictions imposed on the space spanned by the cointegration matrix beta}

\item{\code{df}}{degrees of freedom}

\item{\code{dt}}{timestep of the process}

\item{\code{r2}}{proportion of the variability explained by the model}
}}

