% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{rank.johansen}
\alias{rank.johansen}
\title{Determine the cointegration rank by Johansen procedure}
\usage{
rank.johansen(X, conf.level = 0.05, type = c("trace", "max"))
}
\arguments{
\item{X}{Matrix of lagged values of the process}

\item{conf.level}{confidence level of the test}

\item{type}{which test statistic should be used, the options are
\code{"trace"} (trace test) and \code{"max"} (maximum eigenvalue test).}
}
\value{
A list containing the following components:
}
\description{
This function loads data matrix X and determines the cointegration rank by
likelihood ratio tests proposed by Johansen (trace test or maximum eigenvalue test)
}
\section{Slots}{

\describe{
\item{\code{r}}{cointegration rank}

\item{\code{lambdas}}{eigenvalues}

\item{\code{lrts}}{data frame with test statics for ranks 0 to p-1 (rows) and both types of the test (columns "trace" and "max")}
}}

